+++ src/dialogues/mod.rs
@@ -1,4 +1,4 @@
 pub mod add_expense_dialogue;
-mod keyboard;
+pub(crate) mod keyboard;
 pub mod pending_command_dialogue;
 pub mod storage;